Fix initial implementation of the algorithm

+++ include/CreatedObject.hpp
@@ -2,11 +2,11 @@
 
 namespace tulip::text {
 	struct CreatedObject {
-		float x;
-		float y;
+		double x;
+		double y;
 		
 		int32_t objectId;
-		float scale;
-		float rotation;
+		double scale;
+		double rotation;
 	};
 }
